fix: clippy warnings
- Use repeat_n instead of repeat().take() in fiber.rs
- Remove unnecessary cast in token.rs

+++ crates/zcash-jd-server/src/token.rs
@@ -98,7 +98,7 @@ impl TokenManager {
                 .values()
                 .filter(|t| t.client_id == client_id && !t.is_expired())
                 .count();
-            if client_token_count >= self.config.max_tokens_per_client as usize {
+            if client_token_count >= self.config.max_tokens_per_client {
                 return Err(JdServerError::Protocol(format!(
                     "Max tokens per client exceeded ({})",
                     self.config.max_tokens_per_client
